# services/api/userapi.swagger.yaml
openapi: 3.0.0
info:
  version: "1.0.0"
  title: User Service
servers:
  - description: Default Host URL
    url: http://arch.homework
tags:
  - name: user
    description: Operations about user
paths:
  /api/v1/user:
    get:
      tags:
        - user
      summary: get current user id
      operationId: createUser
      responses:
        '200':
          description: successfull response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserId'
        '403':
          description: forbidden response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/user/{userId}:
    parameters:
      - name: userId
        in: path
        description: ID of user
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags:
        - user
      operationId: getUserProfile
      responses:
        '200':
          description: user response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '403':
          description: forbidden response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - user
      operationId: updateUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
            examples:
              sample-user:
                summary: Example
                value:
                  firstName: Julie
                  lastName: Doe
                  email: bestjohn@doe.com
                  phone: '+71004242424'

      responses:
        '200':
          description: user updated
        '403':
          description: forbidden response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    UserId:
      type: object
      properties:
        id:
          type: string
          format: uuid
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        login:
          type: string
          maxLength: 255
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
          format: phone
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string

  requestBodies:
    UserArray:
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/User'
      description: List of user object
      required: true